refactor(BA-5795): unify delete actions into RBAC-enforced DeleteVFolderV2Action
Remove the separate delete_v2_rbac path and convert the existing
delete_v2 action to use SingleEntityActionProcessor with RBAC
validators, following the same pattern applied to purge_v2 in BA-5765.
