+++ src/ai/backend/manager/services/vfolder/actions/vfolder_v2.py
@@ -1,4 +1,4 @@
-"""V2 vfolder actions — user_id based, no keypair_resource_policy."""
+"""V2 vfolder actions — RBAC-enforced via SingleEntityActionProcessor."""
 
 import uuid
 from dataclasses import dataclass
@@ -13,17 +13,21 @@ from ai.backend.manager.actions.action import BaseActionResult
 from ai.backend.manager.actions.action.scope import BaseScopeAction
 from ai.backend.manager.actions.types import ActionOperationType
 from ai.backend.manager.data.permission.types import RBACElementRef
+from ai.backend.manager.services.vfolder.actions.base import (
+    VFolderSingleEntityAction,
+    VFolderSingleEntityActionResult,
+)
 
 
 @dataclass
-class DeleteVFolderV2Action(BaseScopeAction):
-    user_id: uuid.UUID
+class DeleteVFolderV2Action(VFolderSingleEntityAction):
+    """Soft-delete a vfolder by ID with RBAC enforcement."""
+
     vfolder_id: uuid.UUID
 
     @override
-    @classmethod
-    def entity_type(cls) -> EntityType:
-        return EntityType.VFOLDER
+    def entity_id(self) -> str | None:
+        return str(self.vfolder_id)
 
     @override
     @classmethod
@@ -31,16 +35,8 @@ class DeleteVFolderV2Action(BaseScopeAction):
         return ActionOperationType.DELETE
 
     @override
-    def entity_id(self) -> str | None:
+    def target_entity_id(self) -> str:
         return str(self.vfolder_id)
-
-    @override
-    def scope_type(self) -> ScopeType:
-        return ScopeType.USER
-
-    @override
-    def scope_id(self) -> str:
-        return str(self.user_id)
 
     @override
     def target_element(self) -> RBACElementRef:
@@ -51,11 +47,15 @@ class DeleteVFolderV2Action(BaseScopeAction):
 
 
 @dataclass
-class DeleteVFolderV2ActionResult(BaseActionResult):
+class DeleteVFolderV2ActionResult(VFolderSingleEntityActionResult):
     vfolder_id: uuid.UUID
 
     @override
     def entity_id(self) -> str | None:
+        return str(self.vfolder_id)
+
+    @override
+    def target_entity_id(self) -> str:
         return str(self.vfolder_id)
